requirements.txt
asgiref==3.2.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
datefinder==0.7.0
Django==3.0.4
django-crispy-forms==1.9.0
django-multiselectfield==0.1.12
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9